updated SVM benchmarking
